example/requirements.txt
django>=1.6
django-debug-toolbar
django-staticfiles-select2
django-selectable
django-staticfiles-jquery
django-localflavor
Pillow


